Fix `SymbolRef()` ambiguity on 32-bit platforms

+++ src/Abstract/Symbol.jl
@@ -183,7 +183,7 @@ SymbolRef(symbols::Symbols, sym::SymtabEntry, idx::UInt32)
 Construct a `SymbolRef` object pointing to the given `SymtabEntry`, which
 itself represents the `idx`'th symbol within the given `Symbols` collection.
 """
-@mustimplement SymbolRef(symbols::Symbols, sym::SymtabEntry, idx::UInt32)
+@mustimplement SymbolRef(symbols::Symbols, sym::SymtabEntry, idx)
 @mustimplement deref(sym::SymbolRef)
 
 """
